added user and admin control

+++ TodoApp/routers/todos.py
@@ -4,6 +4,7 @@ from sqlalchemy.orm import Session
 from model import Todos
 from database import SessionLocal
 from pydantic import BaseModel, Field
+from .auth import get_current_user
 
 router = APIRouter(
     prefix='/todos',
@@ -18,6 +19,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -27,26 +29,34 @@ class TodoRequest(BaseModel):
     
 
 @router.get('/')
-async def read_todos(db: db_dependency):
-    return db.query(Todos).all()
+async def read_todos(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Unauthorized user')
+    return db.query(Todos).filter(Todos.owner == user.get('id')).all()
 
 @router.get('/todos/{todo_id}', status_code=status.HTTP_200_OK)
-async def get_todo_by_id(db: db_dependency, todo_id: int = Path(gt = 0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def get_todo_by_id(user: user_dependency, db: db_dependency, todo_id: int = Path(gt = 0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Unauthorized user')
+    todo_model = db.query(Todos).filter(Todos.owner == user.get('id')).filter(Todos.id == todo_id).first()
     if todo_model is not None:
         return todo_model
     else:
         raise HTTPException(status_code=404, detail='Id not found')
 
 @router.post('/createtodo', status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todorequest: TodoRequest):
-    todo_model = Todos(**todorequest.model_dump())
+async def create_todo(user: user_dependency, db: db_dependency, todorequest: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Unauthorized user')
+    todo_model = Todos(**todorequest.model_dump(), owner = user.get('id'))
     db.add(todo_model)
     db.commit()
 
 @router.put('/updatetodo/{todo_id}', status_code=status.HTTP_202_ACCEPTED)
-async def update_todo(db: db_dependency, todorequest: TodoRequest, todo_id: int = Path(gt = 0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency, todorequest: TodoRequest, todo_id: int = Path(gt = 0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Unauthorized user')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Id not found')
     else:
@@ -54,15 +64,16 @@ async def update_todo(db: db_dependency, todorequest: TodoRequest, todo_id: int 
         todo_model.description = todorequest.description
         todo_model.priority = todorequest.priority
         todo_model.completed = todorequest.completed
-
         db.add(todo_model)
         db.commit()
 
 @router.delete('/deletetodo/{todo_id}',status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt = 0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id)
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt = 0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Unauthorized user')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Id not found')
     else:
-        db.query(Todos).filter(Todos.id == todo_id).delete()
+        db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get('id')).delete()
         db.commit()
